+++ node/resolvers/checkout/index.ts
@@ -347,7 +347,7 @@ export const queries: Record<string, Resolver> = {
 
   itemsWithSimulation: async (
     _,
-    { items }: { items: ItemWithSimulationInput[] },
+    { items, regionId }: { items: ItemWithSimulationInput[], regionId?: string },
     ctx: Context
   ) => {
     const {
@@ -359,7 +359,8 @@ export const queries: Record<string, Resolver> = {
       return new Promise((resolve) => {
         const simulationPayloads = getSimulationPayloadsByItem(
           item,
-          segment
+          segment,
+          regionId
         )
 
         const simulationPromises = simulationPayloads.map((payload) =>