+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Use `unitMultiplier` to calculate the price.
+
+### Added
 - Add utmi campaign to the simulation request
+- `regionId` to `itemsWithSimulation` query.
 
 ## [2.144.0] - 2021-08-31
 ### Added