+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.145.0] - 2021-08-31
+
 ### Fixed
 - Use `unitMultiplier` to calculate the price.